+++ CHANGES.rst
@@ -10,7 +10,7 @@ API Changes
 - ``statistic`` argument in ``Background`` initializer can now be either ``"average"``,
   ``"median"``, or a custom callable that takes a ``numpy.ma.MaskedArray`` masked array
   as an input and accepts ``axis`` as an argument. [#253]
-- ``bkg_statistic`` keyword in ``Background.bkg_spectrum()`` now deprecated and raises
+- ``bkg_statistic`` keyword in ``Background.bkg_spectrum()`` is now deprecated and raises
   a warning if set. The ``statistic`` argument in the ``Background`` initializer should
   be used instead. [#253]